HC-3435 Created separate trait for running console commands
